+++ sdk/workloads/azure-mgmt-workloads/azure/mgmt/workloads/aio/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._workloads_client import WorkloadsClient
+from ._workloads_mgmt_client import WorkloadsMgmtClient
 
 try:
     from ._patch import __all__ as _patch_all
@@ -16,7 +16,7 @@ except ImportError:
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
-    "WorkloadsClient",
+    "WorkloadsMgmtClient",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])