+++ sdk/workloads/azure-mgmt-workloads/generated_samples/sap_virtual_instances_create_distributed_skip_transport.py
@@ -7,14 +7,14 @@
 # --------------------------------------------------------------------------
 
 from azure.identity import DefaultAzureCredential
-from azure.mgmt.workloads import WorkloadsClient
+from azure.mgmt.workloads import WorkloadsMgmtClient
 
 """
 # PREREQUISITES
     pip install azure-identity
     pip install azure-mgmt-workloads
 # USAGE
-    python sap_virtual_instances_create_distributed_skip_trans_mount.py
+    python sap_virtual_instances_create_distributed_skip_transport.py
 
     Before run the sample, please set the values of the client ID, tenant ID and client secret
     of the AAD application as environment variables: AZURE_CLIENT_ID, AZURE_TENANT_ID,
@@ -24,7 +24,7 @@ from azure.mgmt.workloads import WorkloadsClient
 
 
 def main():
-    client = WorkloadsClient(
+    client = WorkloadsMgmtClient(
         credential=DefaultAzureCredential(),
         subscription_id="8e17e36c-42e9-4cd5-a078-7b44883414e0",
     )
@@ -36,6 +36,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/workloads/resource-manager/Microsoft.Workloads/preview/2022-11-01-preview/examples/sapvirtualinstances/SAPVirtualInstances_Create_Distributed_SkipTransMount.json
+# x-ms-original-file: specification/workloads/resource-manager/Microsoft.Workloads/stable/2023-04-01/examples/sapvirtualinstances/SAPVirtualInstances_Create_Distributed_SkipTransport.json
 if __name__ == "__main__":
     main()