+++ sdk/workloads/azure-mgmt-workloads/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 1.0.0 (2023-03-03)
+
+### Features Added
+
+  - Added operation group WorkloadsMgmtClientOperationsMixin
+
+### Breaking Changes
+
+  - Client name is changed from `WorkloadsClient` to `WorkloadsMgmtClient`
+  - Model ImageReference no longer has parameter exact_version
+  - Model ImageReference no longer has parameter shared_gallery_image_id
+  - Removed operation group WorkloadsClientOperationsMixin
+
 ## 1.0.0b3 (2023-02-15)
 
 ### Features Added